requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
requests>=2.31.0
python-dotenv>=1.0.0
starlette>=0.28.0
